requirements.txt
numpy
scipy
tk
pillow
